* Fixed bug where HoneypotPreCreateEvent was called twice and HoneypotCreateEvent was never called when a Honeypot was created via the GUI * Fixes a bug where Materials that are not representable as items (Such as REDSTONE_WALL_TORCH) would cause the GUI to throw unhandled exceptions * Fix bug where `getNearbyHoneypots` SQL query made it impossible to return anything.

+++ paper/src/main/java/org/reprogle/honeypot/common/commands/subcommands/HoneypotRemove.java
@@ -97,8 +97,7 @@ public class HoneypotRemove implements HoneypotSubCommand {
     }
 
     private void potRemovalCheck(Block block, Player p) {
-        assert block != null;
-        if (blockManager.isHoneypotBlock(block)) {
+        if (block != null && blockManager.isHoneypotBlock(block)) {
             blockManager.deleteBlock(block);
             p.sendMessage(commandFeedback.sendCommandFeedback("success", false));
         } else {
